bug fix
there was a bug in the possible moves function where the logic to splice out
multiple moves that are not possible was not working as intended.
The splice function was in a for loop and the splice index was being determined by
the index in the loop. If there were multiple entries that needed to be removed splice
was not working because the index would not be correct.